refactored clx module, removed the base driver module and merged it into the clx one.  legacy methods were moved to clx_legacy.LogixDriverLegacy to clean up main module, but not ready to fully remove them yet. added micro800 argument to enable support for Micro800 controllers
