+++ pycomm3/clx.py
@@ -23,24 +23,27 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
+#
 
-import struct
-from collections import defaultdict
-from types import GeneratorType
+import socket
+from functools import wraps
+from os import urandom
 from typing import Union, List, Sequence, Tuple, Optional
 
 from autologging import logged
-from functools import wraps
-from . import DataError, Tag, RequestError
-from .base import Base
-from .bytes_ import (pack_dint, pack_uint, pack_udint, pack_usint, unpack_usint, unpack_uint, unpack_dint, unpack_udint,
-                     pack_sint, UNPACK_DATA_FUNCTION, PACK_DATA_FUNCTION, DATA_FUNCTION_SIZE)
-from .const import (SUCCESS, EXTENDED_SYMBOL, ENCAPSULATION_COMMAND, DATA_TYPE, BITS_PER_INT_TYPE,
-                    REPLY_INFO, TAG_SERVICES_REQUEST, PADDING_BYTE, ELEMENT_ID, DATA_ITEM, ADDRESS_ITEM,
-                    CLASS_ID, CLASS_CODE, INSTANCE_ID, INSUFFICIENT_PACKETS, REPLY_START, BASE_TAG_BIT,
-                    MULTISERVICE_READ_OVERHEAD, MULTISERVICE_WRITE_OVERHEAD, MIN_VER_INSTANCE_IDS, REQUEST_PATH_SIZE,
-                    VENDORS, PRODUCT_TYPES, KEYSWITCH, TAG_SERVICES_REPLY, get_service_status, get_extended_status,
-                    TEMPLATE_MEMBER_INFO_LEN, EXTERNAL_ACCESS, DATA_TYPE_SIZE)
+
+from . import DataError, CommError
+from . import Tag, RequestError
+from .bytes_ import (pack_usint, pack_udint, pack_uint, pack_dint, unpack_uint, unpack_udint, )
+from .bytes_ import (unpack_dint, pack_sint, PACK_DATA_FUNCTION)
+from .const import (DATA_TYPE, TAG_SERVICES_REQUEST, EXTENDED_SYMBOL, PATH_SEGMENTS, ELEMENT_ID, CLASS_CODE, CLASS_ID,
+                    INSTANCE_ID, FORWARD_CLOSE, FORWARD_OPEN, LARGE_FORWARD_OPEN, CONNECTION_MANAGER_INSTANCE, PRIORITY,
+                    TIMEOUT_MULTIPLIER, TIMEOUT_TICKS, TRANSPORT_CLASS, UNCONNECTED_SEND, PRODUCT_TYPES, VENDORS, STATES)
+from .const import (SUCCESS, INSUFFICIENT_PACKETS, BASE_TAG_BIT, MIN_VER_INSTANCE_IDS, REQUEST_PATH_SIZE,
+                    KEYSWITCH, get_service_status, TEMPLATE_MEMBER_INFO_LEN, EXTERNAL_ACCESS, DATA_TYPE_SIZE)
+from .packets import REQUEST_MAP, RequestPacket
+from .socket_ import Socket
+
 
 # re_bit = re.compile(r'(?P<base>^.*)\.(?P<bit>([0-2][0-9])|(3[01])|[0-9])$')
 
@@ -50,7 +53,7 @@ def with_forward_open(func):
 
     @wraps(func)
     def wrapped(self, *args, **kwargs):
-        if not self.forward_open():
+        if not self._forward_open():
             msg = f'Target did not connected. {func.__name__} will not be executed.'
             raise DataError(msg)
         return func(self, *args, **kwargs)
@@ -59,38 +62,92 @@ def with_forward_open(func):
 
 
 @logged
-class LogixDriver(Base):
+class LogixDriver:
     """
     An Ethernet/IP Client library for reading and writing tags in ControlLogix and CompactLogix PLCs.
+    """
 
-    The following services have been implemented:
-        - Read Tag Service (0x4c)
-        - Read Tag Fragment Service (0x52)
-        - Write Tag Service (0x4d)
-        - Write Tag Fragment Service (0x53)
-        - Multiple Service Packet (0x0a)
-        - Read Modify Write Tag (0xce)
-
-"""
-
-    def __init__(self, path: str, *args,  large_packets: bool = True,
+    def __init__(self, path: str, *args,  large_packets: bool = True, debug: bool = False, micro800: bool = False,
                  init_info: bool = True, init_tags: bool = True, init_program_tags: bool = False, **kwargs):
         """
-        :param ip_address: IP address of PLC
-        :param slot: Slot of PLC in chassis (leave at 0 for CompactLogix)
-        :param large_packets: if True, Extended Forward Open will be used (v20+ and EN2T+)
+        :param path: CIP path to intended target
+
+            The path may contain 3 forms:
+
+            - IP Address Only (``10.20.30.100``) - Use for a ControlLogix PLC is in slot 0 or if connecting to a CompactLogix or Micro800 PLC.
+            - IP Address/Slot (``10.20.30.100/1``) - (ControlLogix) if PLC is not in slot 0
+            - CIP Routing Path (``1.2.3.4/backplane/2/enet/6.7.8.9/backplane/0``) - Use if needing to route thru a backplane
+
+            .. note::
+
+                Both the IP Address and IP Address/Slot options are shortcuts, they will be replaced with the
+                CIP path automatically.  The ``enet`` / ``backplane`` (or ``bp``) are symbols for the CIP routing
+                port numbers and will be replaced with the correct value.
+
+        :param large_packets: if True (default), the *Extended Forward Open* service will be used
+
+            .. note::
+
+                *Extended Forward Open* allows the used of 4KBs of service data in each request.
+                The standard *Forward Open* is limited to 500 bytes.  Not all hardware supports the large packet size,
+                like ENET or ENBT modules or ControlLogix version 19 or lower.  You may need to disable this option
+                if forward open requests fail.
+
         :param init_info:  if True, initializes controller info (name, revision, etc) on connect
+
+            .. note::
+
+                Initializing the controller info will enable/disable the use of *Symbol Instance Addressing* in
+                the meth:`.read` and :meth:`.write` methods.  If you disable this option and are using an older firmware
+                (below v21), you will need to set ``plc.use_instance_ids`` to False or the reads and writes will fail.
+
         :param init_tags: if True, uploads all controller-scoped tag definitions on connect
         :param init_program_tags: if True, uploads all program-scoped tag definitions on connect
-        """
-        super().__init__(path, *args, **kwargs)
-        self._cache = None
+        :param debug:  enables certain debugging features, like printing the raw bytes for each packet sent/received
+        :param micro800: set to True if connecting to a Micro800 series PLC, it will disable unsupported features
 
+        .. tip::
+
+            Initialization of tags is required for the :meth:`.read` and :meth:`.write` to work.  This is because
+            they require information about the data type and structure of the tags inside the controller.  If opening
+            multiple connections to the same controller, you may disable tag initialization is all but the first connection
+            and set ``plc2._tags = plc1.tags`` to prevent needing to upload the tag definitions multiple times.
+
+        """
+
+        self._sequence_number = 1
+        self._sock = None
+        # self.__direct_connections = direct_connection
+        self.debug = debug
+        self._micro800 = micro800
+        self._session = 0
+        self._connection_opened = False
+        self._target_cid = None
+        self._target_is_connected = False
+        self._info = {}
+        self.connection_size = 500
+        ip, _path = _parse_connection_path(path, micro800)
+
+        self.attribs = {
+            'context': b'_pycomm_',
+            'protocol version': b'\x01\x00',
+            'rpi': 5000,
+            'port': 0xAF12,  # 44818
+            'timeout': 10,
+            'ip address': ip,
+            'cip_path': _path,
+            'option': 0,
+            'cid': b'\x27\x04\x19\x71',
+            'csn': b'\x27\x04',
+            'vid': b'\x09\x10',
+            'vsn': b'\x09\x10\x19\x71',
+            'name': 'Base',
+            'extended forward open': large_packets}
+        self._cache = None
         self._data_types = {}
         self._program_names = set()
         self._tags = {}
 
-        self.attribs['extended forward open'] = large_packets
         self.connection_size = 4000 if large_packets else 500
         self.use_instance_ids = True
 
@@ -104,6 +161,27 @@ class LogixDriver(Base):
             if init_tags:
                 self.get_tag_list(program='*' if init_program_tags else None)
 
+    def __enter__(self):
+        self.open()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        try:
+            self.close()
+        except CommError:
+            self.__log.exception('Error closing connection.')
+            return False
+        else:
+            if not exc_type:
+                return True
+            else:
+                self.__log.exception('Unhandled Client Error', exc_info=(exc_type, exc_val, exc_tb))
+                return False
+
+    def __repr__(self):
+        _ = self._info
+        return f"Program Name: {_.get('name')}, Device: {_.get('device_type', 'None')}, Revision: {_.get('revision', 'None')}"
+
     @property
     def tags(self) -> dict:
         """
@@ -112,11 +190,322 @@ class LogixDriver(Base):
         return self._tags
 
     @property
-    def data_types(self):
+    def data_types(self) -> dict:
+        """
+        Read-only property for access to all data type definitions uploaded from the controller.
+        """
         return self._data_types
+
+    @property
+    def connected(self) -> bool:
+        """
+        Read-Only Property to check whether or not a connection is open.
+
+        :return: True is a connection is open, False otherwise
+        """
+        return self._connection_opened
+
+    @property
+    def info(self) -> dict:
+        """
+        Property containing a dict of all the information collected about the connected PLC.
+
+        **Fields**:
+
+        - *vendor* - name of hardware vendor, e.g. ``'Rockwell Automation/Allen-Bradley'``
+        - *product_type* - typically ``'Programmable Logic Controller'``
+        - *product_code* - code identifying the product type
+        - *version_major* - numeric value of major firmware version, e.g. ``28``
+        - *version_minor* - numeric value of minor firmware version, e.g ``13``
+        - *revision* - string value of firmware major and minor version, e.g. ``'28.13'``\
+        - *serial* - hex string of PLC serial number, e.g. ``'FFFFFFFF'``
+        - *device_type* - string value for PLC device type, e.g. ``'1756-L83E/B'``
+        - *keyswitch* - string value representing the current keyswitch position, e.g. ``'REMOTE RUN'``
+        - *name* - string value of the current PLC program name, e.g. ``'PLCA'``
+
+        """
+        return self._info
+
+    @property
+    def name(self) -> Optional[str]:
+        """
+        :return: name of PLC program
+        """
+        return self._info.get('name')
+
+    def new_request(self, command: str) -> RequestPacket:
+        """
+        Creates a new request packet for the given command.
+        If the command is invalid, a base :class:`RequestPacket` is created.
+
+        Commands:
+            - `send_unit_data`
+            - `send_rr_data`
+            - `register_session`
+            - `unregister_session`
+            - `list_identity`
+            - `multi_request`
+            - `read_tag`
+            - `read_tag_fragmented`
+            - `write_tag`
+            - `write_tag_fragmented`
+
+        :param command: the service for which a request will be created
+        :return: a new request for the command
+        """
+        cls = REQUEST_MAP[command]
+        return cls(self)
+
+    @property
+    def _sequence(self) -> int:
+        """
+        Increment and return the sequence id used with connected messages
+
+        :return: The next sequence number
+        """
+        self._sequence_number += 1
+
+        if self._sequence_number >= 65535:
+            self._sequence_number = 1
+
+        return self._sequence_number
+
+    def list_identity(self) -> Optional[str]:
+        """
+        Uses the ListIdentity service to identify the target
+
+        :return: device identity if reply contains valid response else None
+        """
+
+        request = self.new_request('list_identity')
+        response = request.send()
+        return response.identity
+
+    def get_module_info(self, slot):
+        try:
+            if not self._forward_open():
+                self.__log.warning("Target did not connected. get_plc_name will not be executed.")
+                raise DataError("Target did not connected. get_plc_name will not be executed.")
+            request = self.new_request('send_rr_data')
+            request.add(
+                # unnconnected send portion
+                UNCONNECTED_SEND,
+                b'\x02',
+                CLASS_ID['8-bit'],
+                b'\x06',  # class
+                INSTANCE_ID["8-bit"],
+                b'\x01',
+                b'\x0A',  # priority
+                b'\x0e\x06\x00',
+
+                # Identity request portion
+                b'\x01',  # Service
+                b'\x02',
+                CLASS_ID['8-bit'],
+                CLASS_CODE['Identity Object'],
+                INSTANCE_ID["8-bit"],
+                b'\x01',  # Instance 1
+                b'\x01\x00',
+                b'\x01',  # backplane
+                pack_usint(slot),
+            )
+            response = request.send()
+
+            if response:
+                info = _parse_identity_object(response.data)
+                return info
+            else:
+                raise DataError(f'send_rr_data did not return valid data - {response.error}')
+
+        except Exception as err:
+            raise DataError(err)
+
+    def open(self):
+        """
+        Creates a new Ethernet/IP socket connection to target device and registers a CIP session.
+
+        :return: True if successful, False otherwise
+        """
+        # handle the socket layer
+        if not self._connection_opened:
+            try:
+                if self._sock is None:
+                    self._sock = Socket()
+                self._sock.connect(self.attribs['ip address'], self.attribs['port'])
+                self._connection_opened = True
+                self.attribs['cid'] = urandom(4)
+                self.attribs['vsn'] = urandom(4)
+                if self._register_session() is None:
+                    self.__log.warning("Session not registered")
+                    return False
+                return True
+            except Exception as e:
+                raise CommError(e)
+
+    def _register_session(self) -> Optional[int]:
+        """
+        Registers a new CIP session with the target.
+
+        :return: the session id if session registered successfully, else None
+        """
+        if self._session:
+            return self._session
+
+        self._session = 0
+        request = self.new_request('register_session')
+        request.add(
+            self.attribs['protocol version'],
+            b'\x00\x00'
+        )
+
+        response = request.send()
+        if response:
+            self._session = response.session
+
+            if self.debug:
+                self.__log.debug(f"Session = {response.session} has been registered.")
+            return self._session
+
+        self.__log.warning('Session has not been registered.')
+        return None
+
+    def _forward_open(self):
+        """
+        Opens a new connection with the target PLC using the *Forward Open* or *Extended Forward Open* service.
+
+        :return: True if connection is open or was successfully opened, False otherwise
+        """
+
+        if self._target_is_connected:
+            return True
+
+        if self._session == 0:
+            raise CommError("A Session Not Registered Before forward_open.")
+
+        init_net_params = (True << 9) | (0 << 10) | (2 << 13) | (False << 15)
+        if self.attribs['extended forward open']:
+            net_params = pack_udint((self.connection_size & 0xFFFF) | init_net_params << 16)
+        else:
+            net_params = pack_uint((self.connection_size & 0x01FF) | init_net_params)
+
+        forward_open_msg = [
+            FORWARD_OPEN if not self.attribs['extended forward open'] else LARGE_FORWARD_OPEN,
+            b'\x02',  # CIP Path size
+            CLASS_ID["8-bit"],  # class type
+            CLASS_CODE["Connection Manager"],  # Volume 1: 5-1
+            INSTANCE_ID["8-bit"],
+            CONNECTION_MANAGER_INSTANCE['Open Request'],
+            PRIORITY,
+            TIMEOUT_TICKS,
+            b'\x00\x00\x00\x00',
+            self.attribs['cid'],
+            self.attribs['csn'],
+            self.attribs['vid'],
+            self.attribs['vsn'],
+            TIMEOUT_MULTIPLIER,
+            b'\x00\x00\x00',
+            b'\x01\x40\x20\x00',
+            net_params,
+            b'\x01\x40\x20\x00',
+            net_params,
+            TRANSPORT_CLASS,
+            self.attribs['cip_path']
+        ]
+        request = self.new_request('send_rr_data')
+        request.add(*forward_open_msg)
+        response = request.send()
+        if response:
+            self._target_cid = response.data[:4]
+            self._target_is_connected = True
+            return True
+        self.__log.warning(f"forward_open failed - {response.error}")
+        return False
+
+    def close(self):
+        """
+        Closes the current connection and un-registers the session.
+        """
+        errs = []
+        try:
+            if self._target_is_connected:
+                self._forward_close()
+            if self._session != 0:
+                self._un_register_session()
+        except Exception as err:
+            errs.append(err)
+            self.__log.warning(f"Error on close() -> session Err: {err}")
+
+        try:
+            if self._sock:
+                self._sock.close()
+        except Exception as err:
+            errs.append(err)
+            self.__log.warning(f"close() -> _sock.close Err: {err}")
+
+        self._sock = None
+        self._target_is_connected = False
+        self._session = 0
+        self._connection_opened = False
+
+        if errs:
+            raise CommError(' - '.join(str(e) for e in errs))
+
+    def _un_register_session(self):
+        """
+        Un-registers the current session with the target.
+        """
+        request = self.new_request('unregister_session')
+        request.send()
+        self._session = None
+
+    def _forward_close(self):
+        """ CIP implementation of the forward close message
+
+        Each connection opened with the forward open message need to be closed.
+        Refer to ODVA documentation Volume 1 3-5.5.3
+
+        :return: False if any error in the replayed message
+        """
+
+        if self._session == 0:
+            raise CommError("A session need to be registered before to call forward_close.")
+        request = self.new_request('send_rr_data')
+
+        path_size, *path = self.attribs['cip_path']  # for some reason we need to add a 0x00 between these? CIP Vol 1
+
+        forward_close_msg = [
+            FORWARD_CLOSE,
+            b'\x02',
+            CLASS_ID["8-bit"],
+            CLASS_CODE["Connection Manager"],  # Volume 1: 5-1
+            INSTANCE_ID["8-bit"],
+            CONNECTION_MANAGER_INSTANCE['Open Request'],
+            PRIORITY,
+            TIMEOUT_TICKS,
+            self.attribs['csn'],
+            self.attribs['vid'],
+            self.attribs['vsn'],
+            bytes([path_size, 0, *path])
+        ]
+
+        request.add(*forward_close_msg)
+        response = request.send()
+        if response:
+            self._target_is_connected = False
+            return True
+
+        self.__log.warning(f"forward_close failed - {response.error}")
+        return False
 
     @with_forward_open
     def get_plc_name(self) -> str:
+        """
+        Requests the name of the program running in the PLC. Uses KB `23341`_ for implementation.
+
+        .. _23341: https://rockwellautomation.custhelp.com/app/answers/answer_view/a_id/23341
+
+        :return:  the controller program name
+        """
         try:
             request = self.new_request('send_unit_data')
             request.add(
@@ -146,19 +535,6 @@ class LogixDriver(Base):
     def get_plc_info(self) -> dict:
         """
         Reads basic information from the controller, returns it and stores it in the ``info`` property.
-
-        info = {
-            'vendor': 'Rockwell Automation/Allen-Bradley',
-            'product_type': 'Programmable Logic Controller',
-            'product_code': 55,
-            'version_major': 20,
-            'version_minor': 12,
-            'revision': '20.12',
-            'serial': '00ff00ff',
-            'device_type': '1756-L62/B LOGIX5562',
-            'keyswitch': 'REMOTE RUN',
-            'name': 'PLCA'
-        }
         """
         try:
             request = self.new_request('send_unit_data')
@@ -186,16 +562,22 @@ class LogixDriver(Base):
     @with_forward_open
     def get_tag_list(self, program: str = None, cache: bool = True) -> List[dict]:
         """
-        Returns the list of tags from the controller. For only controller-scoped tags, set `program` to None (default).
-        Set `program` to a program name to only get the program scoped tags from the specified program.
-        To get all controller and all program scoped tags from all programs, set `program` to '*'
+        Reads the tag list from the controller and the definition for each tag.  Definitions include tag name, tag type
+        (atomic vs struct), data type (including nested definitions for structs), external access, dimensions defined (0-3)
+        for arrays and their length, etc.
 
-        Note, for program scoped tags the tag['tag_name'] will be 'Program:{program}.{tag_name}'. This is so the tag
-        list can be fed directly into the read function.
+        .. note::
 
-        If the `cache` parameter is True (default), the list of tags will be stored so they can be referenced later.
-        This also allows the read/write methods to use the cached instance id's and allow packing more tags into a single
-        request.  While this method returns a list of tags, when cached the tag list is stored as a dict of {tag['tag_name'] : tag, ...}
+            For program scoped tags the tag['tag_name'] will be ``'Program:{program}.{tag_name}'``. This is so the tag
+            list can be fed directly into the read function.
+
+
+        :param program: scope to retrieve tag list, None for controller-only tags, ``'*'`` for all tags, else name of program
+        :param cache: store the retrieved list in the :attr:`.tags` property.  Disable if you wish to get tags retrieved
+                      to not overwrite the currently cached definition. For instance if you're checking tags in a single
+                      program but currently reading controller-scoped tags.
+
+        :return: a list containing dicts for each tag definition collected
         """
 
         self._cache = {
@@ -389,7 +771,7 @@ class LogixDriver(Base):
         """
         if instance_id not in self._cache['id:struct']:
             if not self._target_is_connected:
-                if not self.forward_open():
+                if not self._forward_open():
                     self.__log.warning("Target did not connected. get_tag_list will not be executed.")
                     raise DataError("Target did not connected. get_tag_list will not be executed.")
             request = self.new_request('send_unit_data')
@@ -411,59 +793,11 @@ class LogixDriver(Base):
             response = request.send()
             if not response:
                 raise DataError(f"send_unit_data returned not valid data", response.error)
-            _struct = self._parse_structure_makeup_attributes(response)
+            _struct = _parse_structure_makeup_attributes(response)
             self._cache['id:struct'][instance_id] = _struct
             self._cache['handle:id'][_struct['structure_handle']] = instance_id
 
         return self._cache['id:struct'][instance_id]
-
-    @staticmethod
-    def _parse_structure_makeup_attributes(response):
-        """ extract the tags list from the message received"""
-        structure = {}
-
-        if response.service_status != SUCCESS:
-            structure['Error'] = response.service_status
-            return
-
-        attribute = response.data
-        idx = 4
-        try:
-            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
-                idx += 2
-                structure['object_definition_size'] = unpack_dint(attribute[idx:idx + 4])
-            else:
-                structure['Error'] = 'object_definition Error'
-                return structure
-
-            idx += 6
-            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
-                idx += 2
-                structure['structure_size'] = unpack_dint(attribute[idx:idx + 4])
-            else:
-                structure['Error'] = 'structure Error'
-                return structure
-
-            idx += 6
-            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
-                idx += 2
-                structure['member_count'] = unpack_uint(attribute[idx:idx + 2])
-            else:
-                structure['Error'] = 'member_count Error'
-                return structure
-
-            idx += 4
-            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
-                idx += 2
-                structure['structure_handle'] = unpack_uint(attribute[idx:idx + 2])
-            else:
-                structure['Error'] = 'structure_handle Error'
-                return structure
-
-            return structure
-
-        except Exception as e:
-            raise DataError(e)
 
     def _read_template(self, instance_id, object_definition_size):
         """ get a list of the tags in the plc
@@ -585,6 +919,59 @@ class LogixDriver(Base):
 
         return self._cache['id:udt'][instance_id]
 
+    def _create_tag_rp(self, tag):
+        """
+
+        It returns the request packed wrapped around the tag passed.
+        If any error it returns none
+        """
+        tags = tag.split('.')
+        if tags:
+            base, *attrs = tags
+
+            if self.use_instance_ids and base in self.tags:
+                rp = [CLASS_ID['8-bit'],
+                      CLASS_CODE['Symbol Object'],
+                      INSTANCE_ID['16-bit'], b'\x00',
+                      pack_uint(self.tags[base]['instance_id'])]
+            else:
+                base_tag, index = _find_tag_index(base)
+                base_len = len(base_tag)
+                rp = [EXTENDED_SYMBOL,
+                      pack_usint(base_len),
+                      base_tag]
+                if base_len % 2:
+                    rp.append( b'\x00')
+                if index is None:
+                    return None
+                else:
+                    rp += index
+
+            for attr in attrs:
+                attr, index = _find_tag_index(attr)
+                tag_length = len(attr)
+                # Create the request path
+                attr_path = [EXTENDED_SYMBOL,
+                             pack_usint(tag_length),
+                             attr]
+                # Add pad byte because total length of Request path must be word-aligned
+                if tag_length % 2:
+                    attr_path.append(b'\x00')
+                # Add any index
+                if index is None:
+                    return None
+                else:
+                    attr_path += index
+                rp += attr_path
+
+            # At this point the Request Path is completed,
+            request_path = b''.join(rp)
+            request_path = bytes([len(request_path) // 2]) + request_path
+
+            return request_path
+
+        return None
+
     @with_forward_open
     def read(self, *tags: str) -> Union[Tag, List[Tag]]:
         """
@@ -594,7 +981,7 @@ class LogixDriver(Base):
         """
 
         parsed_requests = self._parse_requested_tags(tags)
-        requests = self._read__build_requests(parsed_requests)
+        requests = self._read_build_requests(parsed_requests)
         read_results = self._send_requests(requests)
 
         results = []
@@ -623,7 +1010,17 @@ class LogixDriver(Base):
         else:
             return results[0]
 
-    def _read__build_requests(self, parsed_tags):
+    def _read_build_requests(self, parsed_tags):
+        if len(parsed_tags) == 1 or self._micro800:
+            requests = (self._read_build_single_request(parsed_tags[tag]) for tag in parsed_tags)
+            return [r for r in requests if r is not None]
+        else:
+            return self._read_build_multi_requests(parsed_tags)
+
+    def _read_build_multi_requests(self, parsed_tags):
+        """
+        creates a list of multi-request packets
+        """
         requests = []
         response_size = 0
         current_request = self.new_request('multi_request')
@@ -640,14 +1037,12 @@ class LogixDriver(Base):
                 else:
                     try:
                         if response_size + return_size < self.connection_size:
-                            if current_request.add_read(tag_data['plc_tag'], tag_data['elements'],
-                                                        tag_data['tag_info']):
+                            if current_request.add_read(tag_data['plc_tag'], tag_data['elements'], tag_data['tag_info']):
                                 response_size += return_size
                             else:
                                 response_size = return_size
                                 current_request = self.new_request('multi_request')
-                                current_request.add_read(tag_data['plc_tag'], tag_data['elements'],
-                                                         tag_data['tag_info'])
+                                current_request.add_read(tag_data['plc_tag'], tag_data['elements'], tag_data['tag_info'])
                                 requests.append(current_request)
                         else:
                             response_size = return_size
@@ -657,8 +1052,30 @@ class LogixDriver(Base):
                     except RequestError:
                         self.__log.exception(f'Failed to build request for {tag} - skipping')
                         continue
+            else:
+                self.__log.error(f'Skipping making request for {tag}, error: {tag_data.get("error")}')
+                continue
 
         return requests
+
+    def _read_build_single_request(self, parsed_tag):
+        """
+        creates a single read_tag request packet
+        """
+
+        if parsed_tag.get('error') is None:
+            return_size = _tag_return_size(parsed_tag['tag_info']) * parsed_tag['elements']
+            if return_size > self.connection_size:
+                request = self.new_request('read_tag_fragmented')
+            else:
+                request = self.new_request('read_tag')
+
+            request.add(parsed_tag['plc_tag'], parsed_tag['elements'], parsed_tag['tag_info'])
+
+            return request
+
+        self.__log.error(f'Skipping making request, error: {parsed_tag["error"]}')
+        return None
 
     @with_forward_open
     def write(self, *tags_values: Sequence[Tuple[str, Union[int, float, str, bool]]]) -> Union[Tag, List[Tag]]:
@@ -676,7 +1093,7 @@ class LogixDriver(Base):
             else:
                 bit_tags.add(tag)
 
-        requests, bit_writes = self._write__build_requests(parsed_requests)
+        requests, bit_writes = self._write_build_requests(parsed_requests)
         write_results = self._send_requests(requests)
         results = []
         for tag, value in tags_values:
@@ -700,11 +1117,19 @@ class LogixDriver(Base):
         else:
             return results[0]
 
-    def _write__build_requests(self, parsed_tags):
+
+    def _write_build_requests(self, parsed_tags):
+        bit_writes = {}
+        if len(parsed_tags) == 1 or self._micro800:
+            requests = (self._write_build_single_request(parsed_tags[tag], bit_writes) for tag in parsed_tags)
+            return [r for r in requests if r is not None], bit_writes
+        else:
+            return self._write_build_multi_requests(parsed_tags, bit_writes), bit_writes
+
+    def _write_build_multi_requests(self, parsed_tags, bit_writes):
         requests = []
         current_request = self.new_request('multi_request')
         requests.append(current_request)
-        bit_writes = {}
 
         tags_in_requests = set()
         for tag, tag_data in parsed_tags.items():
@@ -750,7 +1175,40 @@ class LogixDriver(Base):
                 except RequestError:
                     self.__log.exception(f'Failed to build request for {tag} - skipping')
                     continue
-        return requests, bit_writes
+
+        return requests
+
+    def _write_build_single_request(self, parsed_tag, bit_writes):
+        if parsed_tag.get('error') is None:
+
+            string = _make_string_bytes(parsed_tag)
+            if string is not None:
+                parsed_tag['value'] = string
+
+            if not _bit_request(parsed_tag, bit_writes):
+                parsed_tag['write_value'] = writable_value(parsed_tag['value'], parsed_tag['elements'],
+                                                         parsed_tag['tag_info']['data_type'])
+                if len(parsed_tag['write_value']) > self.connection_size:
+                    request = self.new_request('write_tag_fragmented')
+                else:
+                    request = self.new_request('write_tag')
+
+                request.add(parsed_tag['plc_tag'], parsed_tag['write_value'], parsed_tag['elements'],
+                            parsed_tag['tag_info'])
+                return request
+            else:
+                try:
+                    tag = parsed_tag['plc_tag']
+                    value = bit_writes[tag]['or_mask'], bit_writes[tag]['and_mask']
+                    request = self.new_request('write_tag')
+                    request.add(tag, value, tag_info=bit_writes[tag]['tag_info'], bits_write=True)
+                    return request
+                except RequestError:
+                    self.__log.exception(f'Failed to build request for {tag} - skipping')
+                    return None
+        else:
+            self.__log.error(f'Skipping making request, error: {parsed_tag["error"]}')
+            return None
 
     def _get_tag_info(self, base, attrs) -> Optional[dict]:
 
@@ -829,7 +1287,7 @@ class LogixDriver(Base):
 
         except Exception as err:
             # something went wrong parsing the tag path
-            raise RequestError('Failed to parse tag read request', tag)
+            raise RequestError('Failed to parse tag request', tag)
 
     @staticmethod
     def _send_requests(requests):
@@ -868,665 +1326,6 @@ class LogixDriver(Base):
                                                          tag.get('error', 'Unknown Service Error'))
         return results
 
-    # --------------------------------------------------------------
-    #  OLD CODE - to be removed
-    #
-    # --------------------------------------------------------------
-
-    def create_tag_rp(self, tag):
-        """ Creates a request pad
-
-        It returns the request packed wrapped around the tag passed.
-        If any error it returns none
-        """
-        tags = tag.split('.')
-        if tags:
-            base, *attrs = tags
-
-            if self.use_instance_ids and base in self.tags:
-                rp = [CLASS_ID['8-bit'],
-                      CLASS_CODE['Symbol Object'],
-                      INSTANCE_ID['16-bit'], b'\x00',
-                      pack_uint(self.tags[base]['instance_id'])]
-            else:
-                base_tag, index = self._find_tag_index(base)
-                base_len = len(base_tag)
-                rp = [EXTENDED_SYMBOL,
-                      pack_usint(base_len),
-                      base_tag]
-                if base_len % 2:
-                    rp.append(PADDING_BYTE)
-                if index is None:
-                    return None
-                else:
-                    rp += index
-
-            for attr in attrs:
-                attr, index = self._find_tag_index(attr)
-                tag_length = len(attr)
-                # Create the request path
-                attr_path = [EXTENDED_SYMBOL,
-                             pack_usint(tag_length),
-                             attr]
-                # Add pad byte because total length of Request path must be word-aligned
-                if tag_length % 2:
-                    attr_path.append(PADDING_BYTE)
-                # Add any index
-                if index is None:
-                    return None
-                else:
-                    attr_path += index
-                rp += attr_path
-
-            # At this point the Request Path is completed,
-            request_path = b''.join(rp)
-            request_path = bytes([len(request_path) // 2]) + request_path
-
-            return request_path
-
-        return None
-
-    def _find_tag_index(self, tag):
-        if '[' in tag:  # Check if is an array tag
-            t = tag[:len(tag) - 1]  # Remove the last square bracket
-            inside_value = t[t.find('[') + 1:]  # Isolate the value inside bracket
-            index = inside_value.split(',')  # Now split the inside value in case part of multidimensional array
-            tag = t[:t.find('[')]  # Get only the tag part
-        else:
-            index = []
-        return tag.encode(), self._encode_tag_index(index)
-
-    @staticmethod
-    def _encode_tag_index(index):
-        path = []
-        for idx in index:
-            val = int(idx)
-            if val <= 0xff:
-                path += [ELEMENT_ID["8-bit"], pack_usint(val)]
-            elif val <= 0xffff:
-                path += [ELEMENT_ID["16-bit"], PADDING_BYTE, pack_uint(val)]
-            elif val <= 0xfffffffff:
-                path += [ELEMENT_ID["32-bit"], PADDING_BYTE, pack_dint(val)]
-            else:
-                return None  # Cannot create a valid request packet
-        return path
-
-    def _check_reply(self, reply):
-        """ check the replayed message for error
-
-            return the status error if unsuccessful, else None
-        """
-        try:
-            if reply is None:
-                return f'{REPLY_INFO[unpack_dint(reply[:2])]} without reply'
-            # Get the type of command
-            typ = unpack_uint(reply[:2])
-
-            # Encapsulation status check
-            if unpack_dint(reply[8:12]) != SUCCESS:
-                return get_service_status(unpack_dint(reply[8:12]))
-
-            # Command Specific Status check
-            if typ == unpack_uint(ENCAPSULATION_COMMAND["send_rr_data"]):
-                status = unpack_usint(reply[42:43])
-                if status != SUCCESS:
-                    return f"send_rr_data reply:{get_service_status(status)} - " \
-                           f"Extend status:{get_extended_status(reply, 42)}"
-                else:
-                    return None
-            elif typ == unpack_uint(ENCAPSULATION_COMMAND["send_unit_data"]):
-                service = reply[46]
-                status = _unit_data_status(reply)
-                # return None
-                if status == INSUFFICIENT_PACKETS and service in (TAG_SERVICES_REPLY['Read Tag'],
-                                                                  TAG_SERVICES_REPLY['Multiple Service Packet'],
-                                                                  TAG_SERVICES_REPLY['Read Tag Fragmented'],
-                                                                  TAG_SERVICES_REPLY['Write Tag Fragmented'],
-                                                                  TAG_SERVICES_REPLY['Get Instance Attributes List'],
-                                                                  TAG_SERVICES_REPLY['Get Attributes']):
-                    return None
-                if status == SUCCESS:
-                    return None
-
-                return f"send_unit_data reply:{get_service_status(status)} - " \
-                       f"Extend status:{get_extended_status(reply, 48)}"
-
-        except Exception as e:
-            raise DataError(e)
-
-    def read_tag(self, *tags):
-        """ read tag from a connected plc
-
-        Possible combination can be passed to this method:
-                - ('Counts') a single tag name
-                - (['ControlWord']) a list with one tag or many
-                - (['parts', 'ControlWord', 'Counts'])
-
-        At the moment there is not a strong validation for the argument passed. The user should verify
-        the correctness of the format passed.
-
-        :return: None is returned in case of error otherwise the tag list is returned
-        """
-
-        if not self.forward_open():
-            self.__log.warning("Target did not connected. read_tag will not be executed.")
-            raise DataError("Target did not connected. read_tag will not be executed.")
-
-        if len(tags) == 1:
-            if isinstance(tags[0], (list, tuple, GeneratorType)):
-                return self._read_tag_multi(tags[0])
-            else:
-                return self._read_tag_single(tags[0])
-        else:
-            return self._read_tag_multi(tags)
-
-    def _read_tag_multi(self, tags):
-        tag_bits = defaultdict(list)
-        rp_list, tags_read = [[]], [[]]
-        request_len = 0
-        for tag in tags:
-            tag, bit = self._prep_bools(tag, 'BOOL', bits_only=True)
-            read = bit is None or tag not in tag_bits
-            if bit is not None:
-                tag_bits[tag].append(bit)
-            if read:
-                rp = self.create_tag_rp(tag)
-                if rp is None:
-                    raise DataError(f"Cannot create tag {tag} request packet. read_tag will not be executed.")
-                else:
-                    tag_req_len = len(rp) + MULTISERVICE_READ_OVERHEAD
-                    if tag_req_len + request_len >= self.connection_size:
-                        rp_list.append([])
-                        tags_read.append([])
-                        request_len = 0
-                    rp_list[-1].append(bytes([TAG_SERVICES_REQUEST['Read Tag']]) + rp + b'\x01\x00')
-                    tags_read[-1].append(tag)
-                    request_len += tag_req_len
-
-        replies = []
-        for req_list, tags_ in zip(rp_list, tags_read):
-            message_request = self.build_multiple_service(req_list, self._get_sequence())
-            msg = self.build_common_packet_format(DATA_ITEM['Connected'], b''.join(message_request),
-                                                  ADDRESS_ITEM['Connection Based'], addr_data=self._target_cid, )
-            print(msg)
-            success, reply = self.send_unit_data(msg)
-            if not success:
-                raise DataError(f"send_unit_data returned not valid data - {reply}")
-
-            replies += self._parse_multiple_request_read(reply, tags_, tag_bits)
-        return replies
-
-    def _read_tag_single(self, tag):
-        tag, bit = self._prep_bools(tag, 'BOOL', bits_only=True)
-        rp = self.create_tag_rp(tag)
-        if rp is None:
-            self.__log.warning(f"Cannot create tag {tag} request packet. read_tag will not be executed.")
-            return None
-        else:
-            # Creating the Message Request Packet
-            message_request = [
-                pack_uint(self._get_sequence()),
-                bytes([TAG_SERVICES_REQUEST['Read Tag']]),  # the Request Service
-                # bytes([len(rp) // 2]),  # the Request Path Size length in word
-                rp,  # the request path
-                b'\x01\x00',
-            ]
-        request = self.build_common_packet_format(DATA_ITEM['Connected'], b''.join(message_request),
-                                                  ADDRESS_ITEM['Connection Based'], addr_data=self._target_cid, )
-        success, reply = self.send_unit_data(request)
-
-        if success:
-            data_type = unpack_uint(reply[50:52])
-            typ = DATA_TYPE[data_type]
-            try:
-                value = UNPACK_DATA_FUNCTION[typ](reply[52:])
-                if bit is not None:
-                    value = bool(value & (1 << bit)) if bit < BITS_PER_INT_TYPE[typ] else None
-                return Tag(tag, value, typ)
-            except Exception as e:
-                raise DataError(e)
-        else:
-            return Tag(tag, None, None, reply)
-
-    @staticmethod
-    def _parse_multiple_request_read(reply, tags, tag_bits=None):
-        """ parse the message received from a multi request read:
-
-        For each tag parsed, the information extracted includes the tag name, the value read and the data type.
-        Those information are appended to the tag list as tuple
-
-        :return: the tag list
-        """
-        offset = 50
-        position = 50
-        tag_bits = tag_bits or {}
-        try:
-            number_of_service_replies = unpack_uint(reply[offset:offset + 2])
-            tag_list = []
-            for index in range(number_of_service_replies):
-                position += 2
-                start = offset + unpack_uint(reply[position:position + 2])
-                general_status = unpack_usint(reply[start + 2:start + 3])
-                tag = tags[index]
-                if general_status == SUCCESS:
-                    typ = DATA_TYPE[unpack_uint(reply[start + 4:start + 6])]
-                    value_begin = start + 6
-                    value_end = value_begin + DATA_FUNCTION_SIZE[typ]
-                    value = UNPACK_DATA_FUNCTION[typ](reply[value_begin:value_end])
-                    if tag in tag_bits:
-                        for bit in tag_bits[tag]:
-                            val = bool(value & (1 << bit)) if bit < BITS_PER_INT_TYPE[typ] else None
-                            tag_list.append(Tag(f'{tag}.{bit}', val, 'BOOL'))
-                    else:
-                        tag_list.append(Tag(tag, value, typ))
-                else:
-                    tag_list.append(Tag(tag, None, None, get_service_status(general_status)))
-
-            return tag_list
-        except Exception as e:
-            raise DataError(e)
-
-    def read_array(self, tag, counts, raw=False):
-        """ read array of atomic data type from a connected plc
-
-        At the moment there is not a strong validation for the argument passed. The user should verify
-        the correctness of the format passed.
-
-        :param tag: the name of the tag to read
-        :param counts: the number of element to read
-        :param raw: the value should output as raw-value (hex)
-        :return: None is returned in case of error otherwise the tag list is returned
-        """
-
-        if not self._target_is_connected:
-            if not self.forward_open():
-                self.__log.warning("Target did not connected. read_tag will not be executed.")
-                raise DataError("Target did not connected. read_tag will not be executed.")
-
-        offset = 0
-        last_idx = 0
-        tags = b'' if raw else []
-
-        while offset != -1:
-            rp = self.create_tag_rp(tag)
-            if rp is None:
-                self.__log.warning(f"Cannot create tag {tag} request packet. read_tag will not be executed.")
-                return None
-            else:
-                # Creating the Message Request Packet
-                message_request = [
-                    pack_uint(self._get_sequence()),
-                    bytes([TAG_SERVICES_REQUEST["Read Tag Fragmented"]]),  # the Request Service
-                    # bytes([len(rp) // 2]),  # the Request Path Size length in word
-                    rp,  # the request path
-                    pack_uint(counts),
-                    pack_dint(offset)
-                ]
-            msg = self.build_common_packet_format(DATA_ITEM['Connected'],
-                                                  b''.join(message_request),
-                                                  ADDRESS_ITEM['Connection Based'],
-                                                  addr_data=self._target_cid, )
-            success, reply = self.send_unit_data(msg)
-            if not success:
-                raise DataError(f"send_unit_data returned not valid data - {reply}")
-
-            last_idx, offset = self._parse_fragment(reply, last_idx, offset, tags, raw)
-
-        return tags
-
-    def _parse_fragment(self, reply, last_idx, offset, tags, raw=False):
-        """ parse the fragment returned by a fragment service."""
-
-        try:
-            status = _unit_data_status(reply)
-            data_type = unpack_uint(reply[REPLY_START:REPLY_START + 2])
-            fragment_returned = reply[REPLY_START + 2:]
-        except Exception as e:
-            raise DataError(e)
-
-        fragment_returned_length = len(fragment_returned)
-        idx = 0
-        while idx < fragment_returned_length:
-            try:
-                typ = DATA_TYPE[data_type]
-                if raw:
-                    value = fragment_returned[idx:idx + DATA_FUNCTION_SIZE[typ]]
-                else:
-                    value = UNPACK_DATA_FUNCTION[typ](fragment_returned[idx:idx + DATA_FUNCTION_SIZE[typ]])
-                idx += DATA_FUNCTION_SIZE[typ]
-            except Exception as e:
-                raise DataError(e)
-            if raw:
-                tags += value
-            else:
-                tags.append((last_idx, value))
-                last_idx += 1
-
-        if status == SUCCESS:
-            offset = -1
-        elif status == 0x06:
-            offset += fragment_returned_length
-        else:
-            self.__log.warning('{0}: {1}'.format(get_service_status(status), get_extended_status(reply, 48)))
-            offset = -1
-
-        return last_idx, offset
-
-    @staticmethod
-    def _prep_bools(tag, typ, bits_only=True):
-        """
-        if tag is a bool and a bit of an integer, returns the base tag and the bit value,
-        else returns the tag name and None
-
-        """
-        if typ != 'BOOL':
-            return tag, None
-        if not bits_only and tag.endswith(']'):
-            try:
-                base, idx = tag[:-1].rsplit(sep='[', maxsplit=1)
-                idx = int(idx)
-                base = f'{base}[{idx // 32}]'
-                return base, idx
-            except Exception:
-                return tag, None
-        else:
-            try:
-                base, bit = tag.rsplit('.', maxsplit=1)
-                bit = int(bit)
-                return base, bit
-            except Exception:
-                return tag, None
-
-    @staticmethod
-    def _dword_to_boolarray(tag, bit):
-        base, tmp = tag.rsplit(sep='[', maxsplit=1)
-        i = int(tmp[:-1])
-        return f'{base}[{(i * 32) + bit}]'
-
-    def _write_tag_multi_write(self, tags):
-        rp_list = [[]]
-        tags_added = [[]]
-        request_len = 0
-        for name, value, typ in tags:
-            name, bit = self._prep_bools(name, typ, bits_only=False)  # check if bool & if bit of int or bool array
-            # Create the request path to wrap the tag name
-            rp = self.create_tag_rp(name, multi_requests=True)
-            if rp is None:
-                self.__log.warning(f"Cannot create tag {tags} req. packet. write_tag will not be executed")
-                return None
-            else:
-                try:
-                    if bit is not None:  # then it is a boolean array
-                        rp = self.create_tag_rp(name, multi_requests=True)
-                        request = bytes([TAG_SERVICES_REQUEST["Read Modify Write Tag"]]) + rp
-                        request += b''.join(self._make_write_bit_data(bit, value, bool_ary='[' in name))
-                        if typ == 'BOOL' and name.endswith(']'):
-                            name = self._dword_to_boolarray(name, bit)
-                        else:
-                            name = f'{name}.{bit}'
-                    else:
-                        request = (bytes([TAG_SERVICES_REQUEST["Write Tag"]]) +
-                                   rp +
-                                   pack_uint(DATA_TYPE[typ]) +
-                                   b'\x01\x00' +
-                                   PACK_DATA_FUNCTION[typ](value))
-
-                    tag_req_len = len(request) + MULTISERVICE_WRITE_OVERHEAD
-                    if tag_req_len + request_len >= self.connection_size:
-                        rp_list.append([])
-                        tags_added.append([])
-                        request_len = 0
-                    rp_list[-1].append(request)
-                    request_len += tag_req_len
-                except (LookupError, struct.error) as e:
-                    self.__warning(f"Tag:{name} type:{typ} removed from write list. Error:{e}.")
-
-                    # The tag in idx position need to be removed from the rp list because has some kind of error
-                else:
-                    tags_added[-1].append((name, value, typ))
-
-        # Create the message request
-        replies = []
-        for req_list, tags_ in zip(rp_list, tags_added):
-            message_request = self.build_multiple_service(req_list, self._get_sequence())
-            msg = self.build_common_packet_format(DATA_ITEM['Connected'],
-                                                  b''.join(message_request),
-                                                  ADDRESS_ITEM['Connection Based'],
-                                                  addr_data=self._target_cid, )
-            success, reply = self.send_unit_data(msg)
-            if success:
-                replies += self._parse_multiple_request_write(tags_, reply)
-            else:
-                raise DataError(f"send_unit_data returned not valid data - {reply}")
-        return replies
-
-    def _write_tag_single_write(self, tag, value, typ):
-        name, bit = self._prep_bools(tag, typ,
-                                     bits_only=False)  # check if we're writing a bit of a integer rather than a BOOL
-
-        rp = self.create_tag_rp(name)
-        if rp is None:
-            self.__log.warning(f"Cannot create tag {tag} request packet. write_tag will not be executed.")
-            return None
-        else:
-            # Creating the Message Request Packet
-            message_request = [
-                pack_uint(self._get_sequence()),
-                bytes([TAG_SERVICES_REQUEST["Read Modify Write Tag"]
-                       if bit is not None else TAG_SERVICES_REQUEST["Write Tag"]]),
-                # bytes([len(rp) // 2]),  # the Request Path Size length in word
-                rp,  # the request path
-            ]
-            if bit is not None:
-                try:
-                    message_request += self._make_write_bit_data(bit, value, bool_ary='[' in name)
-                except Exception as err:
-                    raise DataError(f'Unable to write bit, invalid bit number {repr(err)}')
-            else:
-                message_request += [
-                    pack_uint(DATA_TYPE[typ]),  # data type
-                    pack_uint(1),  # Add the number of tag to write
-                    PACK_DATA_FUNCTION[typ](value)
-                ]
-            request = self.build_common_packet_format(DATA_ITEM['Connected'], b''.join(message_request),
-                                                      ADDRESS_ITEM['Connection Based'], addr_data=self._target_cid)
-            success, reply = self.send_unit_data(request)
-            return Tag(tag, value, typ, None if success else reply)
-
-    @staticmethod
-    def _make_write_bit_data(bit, value, bool_ary=False):
-        or_mask, and_mask = 0x00000000, 0xFFFFFFFF
-
-        if bool_ary:
-            mask_size = 4
-            bit = bit % 32
-        else:
-            mask_size = 1 if bit < 8 else 2 if bit < 16 else 4
-
-        if value:
-            or_mask |= (1 << bit)
-        else:
-            and_mask &= ~(1 << bit)
-
-        return [pack_uint(mask_size), pack_udint(or_mask)[:mask_size], pack_udint(and_mask)[:mask_size]]
-
-    @staticmethod
-    def _parse_multiple_request_write(tags, reply):
-        """ parse the message received from a multi request writ:
-
-        For each tag parsed, the information extracted includes the tag name and the status of the writing.
-        Those information are appended to the tag list as tuple
-
-        :return: the tag list
-        """
-        offset = 50
-        position = 50
-
-        try:
-            number_of_service_replies = unpack_uint(reply[offset:offset + 2])
-            tag_list = []
-            for index in range(number_of_service_replies):
-                position += 2
-                start = offset + unpack_uint(reply[position:position + 2])
-                general_status = unpack_usint(reply[start + 2:start + 3])
-                error = None if general_status == SUCCESS else get_service_status(general_status)
-                tag_list.append(Tag(*tags[index], error))
-            return tag_list
-        except Exception as e:
-            raise DataError(e)
-
-    def write_tag(self, tag, value=None, typ=None):
-        """ write tag/tags from a connected plc
-
-        Possible combination can be passed to this method:
-                - ('tag name', Value, data type)  as single parameters or inside a tuple
-                - ([('tag name', Value, data type), ('tag name2', Value, data type)]) as array of tuples
-
-        At the moment there is not a strong validation for the argument passed. The user should verify
-        the correctness of the format passed.
-
-        The type accepted are:
-            - BOOL
-            - SINT
-            - INT
-            - DINT
-            - REAL
-            - LINT
-            - BYTE
-            - WORD
-            - DWORD
-            - LWORD
-
-        :param tag: tag name, or an array of tuple containing (tag name, value, data type)
-        :param value: the value to write or none if tag is an array of tuple or a tuple
-        :param typ: the type of the tag to write or none if tag is an array of tuple or a tuple
-        :return: None is returned in case of error otherwise the tag list is returned
-        """
-
-        if not self._target_is_connected:
-            if not self.forward_open():
-                self.__log.warning("Target did not connected. write_tag will not be executed.")
-                raise DataError("Target did not connected. write_tag will not be executed.")
-
-        if isinstance(tag, (list, tuple, GeneratorType)):
-            return self._write_tag_multi_write(tag)
-        else:
-            if isinstance(tag, tuple):
-                name, value, typ = tag
-            else:
-                name = tag
-            return self._write_tag_single_write(name, value, typ)
-
-    def write_array(self, tag, values, data_type, raw=False):
-        """ write array of atomic data type from a connected plc
-        At the moment there is not a strong validation for the argument passed. The user should verify
-        the correctness of the format passed.
-        :param tag: the name of the tag to read
-        :param data_type: the type of tag to write
-        :param values: the array of values to write, if raw: the frame with bytes
-        :param raw: indicates that the values are given as raw values (hex)
-        """
-
-        if not isinstance(values, list):
-            self.__log.warning("A list of tags must be passed to write_array.")
-            raise DataError("A list of tags must be passed to write_array.")
-
-        if not self._target_is_connected:
-            if not self.forward_open():
-                self.__log.warning("Target did not connected. write_array will not be executed.")
-                raise DataError("Target did not connected. write_array will not be executed.")
-
-        array_of_values = b''
-        byte_size = 0
-        byte_offset = 0
-
-        for i, value in enumerate(values):
-            array_of_values += value if raw else PACK_DATA_FUNCTION[data_type](value)
-            byte_size += DATA_FUNCTION_SIZE[data_type]
-
-            if byte_size >= 450 or i == len(values) - 1:
-                # create the message and send the fragment
-                rp = self.create_tag_rp(tag)
-                if rp is None:
-                    self.__log.warning(f"Cannot create tag {tag} request packet write_array will not be executed.")
-                    return None
-                else:
-                    # Creating the Message Request Packet
-                    message_request = [
-                        pack_uint(self._get_sequence()),
-                        bytes([TAG_SERVICES_REQUEST["Write Tag Fragmented"]]),  # the Request Service
-                        bytes([len(rp) // 2]),  # the Request Path Size length in word
-                        rp,  # the request path
-                        pack_uint(DATA_TYPE[data_type]),  # Data type to write
-                        pack_uint(len(values)),  # Number of elements to write
-                        pack_dint(byte_offset),
-                        array_of_values  # Fragment of elements to write
-                    ]
-                    byte_offset += byte_size
-
-                msg = self.build_common_packet_format(
-                    DATA_ITEM['Connected'],
-                    b''.join(message_request),
-                    ADDRESS_ITEM['Connection Based'],
-                    addr_data=self._target_cid,
-                )
-
-                success, reply = self.send_unit_data(msg)
-                if not success:
-                    raise DataError(f"send_unit_data returned not valid data - {reply}")
-
-                array_of_values = b''
-                byte_size = 0
-        return True
-
-    def write_string(self, tag, value, size=82):
-        """
-            Rockwell define different string size:
-                STRING  STRING_12   STRING_16   STRING_20   STRING_40   STRING_8
-            by default we assume size 82 (STRING)
-        """
-        data_tag = ".".join((tag, "DATA"))
-        len_tag = ".".join((tag, "LEN"))
-
-        # create an empty array
-        data_to_send = [0] * size
-        for idx, val in enumerate(value):
-            try:
-                unsigned = ord(val)
-                data_to_send[idx] = unsigned - 256 if unsigned > 127 else unsigned
-            except IndexError:
-                break
-
-        str_len = len(value)
-        if str_len > size:
-            str_len = size
-
-        result_len = self.write_tag(len_tag, str_len, 'DINT')
-        result_data = self.write_array(data_tag, data_to_send, 'SINT')
-        return result_data and result_len
-
-    def read_string(self, tag, str_len=None):
-        data_tag = f'{tag}.DATA'
-        if str_len is None:
-            len_tag = f'{tag}.LEN'
-            tmp = self.read_tag(len_tag)
-            length, _ = tmp or (None, None)
-        else:
-            length = str_len
-
-        if length:
-            values = self.read_array(data_tag, length)
-            if values:
-                _, values = zip(*values)
-                chars = ''.join(chr(v + 256) if v < 0 else chr(v) for v in values)
-                string, *_ = chars.split('\x00', maxsplit=1)
-                return string
-        return None
-
-
-def _unit_data_status(reply):
-    return unpack_usint(reply[48:49])
-
 
 def _parse_plc_name(response):
     if response.service_status != SUCCESS:
@@ -1561,6 +1360,108 @@ def _parse_plc_info(data):
         'device_type': device_type,
         'keyswitch': keyswitch
     }
+
+
+def _parse_identity_object(reply):
+    vendor = unpack_uint(reply[:2])
+    product_type = unpack_uint(reply[2:4])
+    product_code = unpack_uint(reply[4:6])
+    major_fw = int(reply[6])
+    minor_fw = int(reply[7])
+    status = f'{unpack_uint(reply[8:10]):0{16}b}'
+    serial_number = f'{unpack_udint(reply[10:14]):0{8}x}'
+    product_name_len = int(reply[14])
+    tmp = 15 + product_name_len
+    device_type = reply[15:tmp].decode()
+
+    state = unpack_uint(reply[tmp:tmp + 4]) if reply[tmp:] else -1  # some modules don't return a state
+
+    return {
+        'vendor': VENDORS.get(vendor, 'UNKNOWN'),
+        'product_type': PRODUCT_TYPES.get(product_type, 'UNKNOWN'),
+        'product_code': product_code,
+        'version_major': major_fw,
+        'version_minor': minor_fw,
+        'revision': f'{major_fw}.{minor_fw}',
+        'serial': serial_number,
+        'device_type': device_type,
+        'status': status,
+        'state': STATES.get(state, 'UNKNOWN'),
+    }
+
+
+def _find_tag_index(tag):
+    if '[' in tag:  # Check if is an array tag
+        t = tag[:len(tag) - 1]  # Remove the last square bracket
+        inside_value = t[t.find('[') + 1:]  # Isolate the value inside bracket
+        index = inside_value.split(',')  # Now split the inside value in case part of multidimensional array
+        tag = t[:t.find('[')]  # Get only the tag part
+    else:
+        index = []
+    return tag.encode(), _encode_tag_index(index)
+
+
+def _encode_tag_index(index):
+        path = []
+        for idx in index:
+            val = int(idx)
+            if val <= 0xff:
+                path += [ELEMENT_ID["8-bit"], pack_usint(val)]
+            elif val <= 0xffff:
+                path += [ELEMENT_ID["16-bit"], b'\x00', pack_uint(val)]
+            elif val <= 0xfffffffff:
+                path += [ELEMENT_ID["32-bit"], b'\x00', pack_dint(val)]
+            else:
+                return None  # Cannot create a valid request packet
+        return path
+
+
+def _parse_structure_makeup_attributes(response):
+        """ extract the tags list from the message received"""
+        structure = {}
+
+        if response.service_status != SUCCESS:
+            structure['Error'] = response.service_status
+            return
+
+        attribute = response.data
+        idx = 4
+        try:
+            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
+                idx += 2
+                structure['object_definition_size'] = unpack_dint(attribute[idx:idx + 4])
+            else:
+                structure['Error'] = 'object_definition Error'
+                return structure
+
+            idx += 6
+            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
+                idx += 2
+                structure['structure_size'] = unpack_dint(attribute[idx:idx + 4])
+            else:
+                structure['Error'] = 'structure Error'
+                return structure
+
+            idx += 6
+            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
+                idx += 2
+                structure['member_count'] = unpack_uint(attribute[idx:idx + 2])
+            else:
+                structure['Error'] = 'member_count Error'
+                return structure
+
+            idx += 4
+            if unpack_uint(attribute[idx:idx + 2]) == SUCCESS:
+                idx += 2
+                structure['structure_handle'] = unpack_uint(attribute[idx:idx + 2])
+            else:
+                structure['Error'] = 'structure_handle Error'
+                return structure
+
+            return structure
+
+        except Exception as e:
+            raise DataError(e)
 
 
 def writable_value(value, elements, data_type):
@@ -1656,3 +1557,64 @@ def _bit_request(tag_data, bit_requests):
         bits_['and_mask'] &= ~(1 << bit)
 
     return True
+
+
+def _parse_connection_path(path, micro800):
+    ip, *segments = path.split('/')
+    try:
+        socket.inet_aton(ip)
+    except OSError:
+        raise ValueError('Invalid IP Address', ip)
+    segments = [_parse_path_segment(s) for s in segments]
+
+    if not segments:
+        if not micro800:
+            _path = [pack_usint(PATH_SEGMENTS['backplane']), b'\x00']  # default backplane/0
+        else:
+            _path = []
+    elif len(segments) == 1:
+        _path = [pack_usint(PATH_SEGMENTS['backplane']), pack_usint(segments[0])]
+    else:
+        pairs = (segments[i:i + 2] for i in range(0, len(segments), 2))
+        _path = []
+        for port, dest in pairs:
+            if isinstance(dest, bytes):
+                port |= 1 << 4  # set Extended Link Address bit, CIP Vol 1 C-1.3
+                dest_len = len(dest)
+                if dest_len % 2:
+                    dest += b'\x00'
+                _path.extend([pack_usint(port), pack_usint(dest_len), dest])
+            else:
+                _path.extend([pack_usint(port), pack_usint(dest)])
+
+    _path += [
+        CLASS_ID['8-bit'],
+        CLASS_CODE['Message Router'],
+        INSTANCE_ID['8-bit'],
+        b'\x01'
+    ]
+
+    _path_bytes = b''.join(_path)
+
+    if len(_path_bytes) % 2:
+        _path_bytes += b'\x00'
+
+    return ip, pack_usint(len(_path_bytes) // 2) + _path_bytes
+
+
+def _parse_path_segment(segment: str):
+    try:
+        if segment.isnumeric():
+            return int(segment)
+        else:
+            tmp = PATH_SEGMENTS.get(segment.lower())
+            if tmp:
+                return tmp
+            else:
+                try:
+                    socket.inet_aton(segment)
+                    return b''.join(pack_usint(ord(c)) for c in segment)
+                except OSError:
+                    raise ValueError('Invalid IP Address Segment', segment)
+    except Exception:
+        raise ValueError(f'Failed to parse path segment', segment)
